Create battery_health_codechef
The problem asks us to determine if an iPhone's battery health is optimal based on a threshold of 80%. Given multiple test cases, each with a battery health percentage X, we need to check if 
X is 80 or greater. If so, the battery is in optimal condition, and we output "YES"; otherwise, we output "NO". The solution requires us to process up to 100 test cases, each with battery health values ranging from 0 to 100. The output is case-insensitive, meaning any combination of upper or lowercase letters for "YES" and "NO" is acceptable.
Problrm link: https://www.codechef.com/problems/BTRYHLTH?tab=statement