parity_codechef
The problem asks to determine whether Ashu and Arvind can equally divide a given number of chocolates, N, for multiple test cases. For each test case, the number of chocolates N is provided, and the task is to check if N is divisible by 2, since only an even number can be divided equally without breaking a chocolate. If N is even, the output should be "Yes", indicating equal division is possible, otherwise "No". This needs to be done for T test cases, and the output for each test case can be any variation of "Yes" or "No" as capitalization doesn't matter. Problem link: https://www.codechef.com/problems/PAR2